move query runner component tests
